Add validation and rejection for recipes (#56)
* 🐛 change code to user username to log in and not email

* ➕ Add security measures to avoid deletion and modification by understanding the routing system

* 🐛 add missing return

* ➕ Add request to get recipes_stash

* ➕ Add recipe in stash handling when user is editor

* ➕ Add display for recipes to confirm

* ➕ Add possibility to reject a stash

* ➕ Add possibility to validate new recipes

* ➕ Add validation for edition

+++ assets/php/controllers/controller.php
@@ -18,14 +18,15 @@ function getAllRecipes($textSupp="") {
     require('./assets/php/views/allRecipesView.php');
 }
 
-function buildRecipeDisplayAllRecipes(&$content, $recipes) {
+function buildRecipeDisplayAllRecipes(&$content, $recipes, $isStash=false) {
     $optionalButtons = "";
-    if (isset($_SESSION["userType"]) && $_SESSION["userType"] = "ADMINISTRATEUR") {
-        $optionalButtons .= "<button><a href='index.php?action=recipeEdition&value=RECI_ID'>Modification</a></button>";
-        $optionalButtons .= "<button><a href='index.php?action=recipeDeletion&value=RECI_ID'>Suppression</a></button>";
-    }
+    $action = $isStash ? "recipeStash" : "recipe";
     for ($i= 0; $i < count($recipes); $i++){
         $recipe = $recipes[$i];
+        if (!$isStash && checkCanEditOrDelete($recipe["users_nickname"])) {
+            $optionalButtons .= "<button><a href='index.php?action=recipeEdition&value=RECI_ID'>Modification</a></button>";
+            $optionalButtons .= "<button><a href='index.php?action=recipeDeletion&value=RECI_ID'>Suppression</a></button>";
+        }
         $recipe_id = $recipe['reci_id'];
         $title = $recipe['reci_title'];
         $resume = $recipe['reci_resume'];
@@ -33,12 +34,13 @@ function buildRecipeDisplayAllRecipes(&$content, $recipes) {
         $image = $recipe['reci_image'];
         $anchor = $i + 1;
         $content .= "<div id='$anchor'>";
-        $content .= "<img src='$image' alt='image de recette' onclick=\"location.href='index.php?action=recipe&value=$recipe_id'\" width=50px height=50px/>" ;
-        $content .= "<h1 onclick=\"location.href='index.php?action=recipe&value=$recipe_id'\" >$title</h1>";
+        $content .= "<img src='$image' alt='image de recette' onclick=\"location.href='index.php?action=$action&value=$recipe_id'\" width=50px height=50px/>" ;
+        $content .= "<h1 onclick=\"location.href='index.php?action=$action&value=$recipe_id'\" >$title</h1>";
         $content .= "<h2>$type</h2>";
         $content .= "<p>$resume<p>";
         $content .= str_replace("RECI_ID", $recipe_id, $optionalButtons);
-        $content .= "</div>";           
+        $content .= "</div>";
+        $optionalButtons = "";
     }
 }
 
@@ -69,19 +71,8 @@ function welcome() {
     $content .= "<p>$editoText</p></section>";
     require('./assets/php/views/welcomeView.php');
 }
-/*recipe's page controller */
-function recipe($reci_id="") {
-    if ($reci_id===""){
-        if (empty($_GET['value'])) welcome();
-        $reci_id = strip_tags($_GET['value']);
-    }
-    if (!is_numeric($reci_id)) return getAllRecipes();
-    if (intval($reci_id) < 1) return getAllRecipes();
-    $recipe = getOneRecipe($reci_id);
-    if (empty($recipe)) getAllRecipes();
-    $recipe = $recipe[0];
-    $ingredients = getRecipeIngredients($reci_id);
 
+function buildRecipeDisplayOneRecipe($recipe, $ingredients, &$content) {
     // Ingredients building format
     $ingredientsHTML = "";
     if (!empty($ingredients)) {
@@ -103,19 +94,106 @@ function recipe($reci_id="") {
     $creationDate = $recipe['reci_creation_date'];
     $lastUpdateDate = $recipe['reci_edit_date'];
     $editorUsername = $recipe['users_nickname'];
-    $content = "<h1>$title</h1>";
+    $content .= "<h1>$title</h1>";
     $content .= "<p>$type</p>";
     $content .= $ingredientsHTML;
     $content .= "<h2>Recette</h2><p>$reci_content</p>";
     $content .= "<p>Créé le : $creationDate par $editorUsername</p>";
     $content .= "<p>Édité pour la dernière fois le : $lastUpdateDate</p>";
     $content .= "<img src='$image' alt='image de recette' width=200px height=200px/>" ;
-    if (isset($_SESSION["userType"]) && $_SESSION["userType"] = "ADMINISTRATEUR") {
+}
+
+/*recipe's page controller */
+function recipe($reci_id="") {
+    if ($reci_id===""){
+        if (empty($_GET['value'])) welcome();
+        $reci_id = strip_tags($_GET['value']);
+    }
+    if (!is_numeric($reci_id)) return getAllRecipes();
+    if (intval($reci_id) < 1) return getAllRecipes();
+    $recipe = getOneRecipe($reci_id);
+    if (empty($recipe)) getAllRecipes();
+    $recipe = $recipe[0];
+    $ingredients = getRecipeIngredients($reci_id);
+
+    buildRecipeDisplayOneRecipe($recipe, $ingredients, $content);
+    if (checkCanEditOrDelete($recipe['users_nickname'])) {
         $content .= "<button><a href='index.php?action=recipeEdition&value=$reci_id'>Modification</a></button>";
         $content .= "<button><a href='index.php?action=recipeDeletion&value=$reci_id'>Suppression</a></button>";
     }
     require('./assets/php/views/recipeView.php');
 }
+
+/*recipe stash's page controller */
+function recipeStash($reci_stash_id="") {
+    $content = "<h1>Nouvelle recette</h1>";
+    if ($reci_stash_id===""){
+        if (empty($_GET['value'])) welcome();
+        $reci_stash_id = strip_tags($_GET['value']);
+    }
+    if (!is_numeric($reci_stash_id)) return getAllRecipes();
+    if (intval($reci_stash_id) < 1) return getAllRecipes();
+    $recipeStash = getOneRecipeStash($reci_stash_id);
+    if (empty($recipeStash)) getAllRecipes();
+    $recipeStash = $recipeStash[0];
+    $ingredientsStash = getRecipeStashIngredients($reci_stash_id);
+    buildRecipeDisplayOneRecipe($recipeStash, $ingredientsStash, $content);
+    
+
+    $reci_id = $recipeStash["reci_id"];
+    if (is_numeric($reci_id)) {
+        $content .= "<h1>Ancienne version de la recette</h1>";
+        $recipe = getOneRecipe($reci_id);
+        if (empty($recipe)) getAllRecipes();
+        $recipe = $recipe[0];
+        $ingredients = getRecipeIngredients($reci_id);
+        buildRecipeDisplayOneRecipe($recipe, $ingredients, $content);
+    }
+    
+    $content .= "<button><a href='index.php?action=validate&value=$reci_stash_id'>Valider</a></button>";
+    $content .= "<button><a href='index.php?action=refuse&value=$reci_stash_id'>Refuser</a></button>";
+    require('./assets/php/views/recipeView.php');
+}
+
+/* To refuse a stash*/
+function refuse() {
+    if (empty($_GET['value'])) account();
+    $reci_stash_id = strip_tags($_GET['value']);
+    deleteRecipeStash($reci_stash_id);
+    account();
+}
+
+/* To validate a stash */
+function validate() {
+    if (empty($_GET['value'])) account();
+    $reci_stash_id = strip_tags($_GET['value']);
+    $recipeStash = getOneRecipeStash($reci_stash_id);
+    $recipeStash = $recipeStash[0];
+    if ($recipeStash["stash_type_value"] === "CREATION") {
+        $reci_id = createRecipe($recipeStash["reci_title"], $recipeStash["reci_content"],
+        $recipeStash["reci_resume"], $recipeStash["rtype_title"], $recipeStash["reci_image"], $recipeStash["users_nickname"], true);
+        $ingredientsStash = getRecipeStashIngredients($reci_stash_id);
+        $ingredients = array();
+        foreach($ingredientsStash as $key => $value) {
+            array_push($ingredients,$value["ing_title"]);
+        }
+        addRecipesIngredients($reci_id, $ingredients, true);
+        deleteRecipeStash($reci_stash_id);
+        account();
+    } elseif ($recipeStash["stash_type_value"] === "MODIFICATION") {
+        editRecipe($recipeStash["reci_id"], $recipeStash["reci_title"], $recipeStash["reci_content"],
+        $recipeStash["reci_resume"], $recipeStash["rtype_title"], $recipeStash["reci_image"], $recipeStash["users_nickname"], true);
+        $ingredientsStash = getRecipeStashIngredients($reci_stash_id);
+        $ingredients = array();
+        foreach($ingredientsStash as $key => $value) {
+            array_push($ingredients,$value["ing_title"]);
+        }
+        editRecipesIngredients($recipeStash["reci_id"], $ingredients, true);
+        deleteRecipeStash($reci_stash_id);
+        account();
+    }
+}
+
 /*filter's page controller*/
 function filter() {
     $content = "";
@@ -162,11 +240,17 @@ function account() {
     $unlogButton = "";
     if (isset($_SESSION['connected']) && boolval($_SESSION['connected']) === true) {
         $unlogButton = '<button><a href="index.php?action=disconnect">Déconnexion</a></button>';
-        if ($_SESSION['userType'] === "ADMINISTRATEUR") {
+        if (strcmp($_SESSION['userType'],"ADMINISTRATEUR") === 0) {
             $content .= "<h1>Modification de l'édito</h1>";
             $content .= "<form id='form_edito' name='edito' method='post' action='index.php?action=editoEdition'>";
             $content .= "<label for='edito'>Entrez le nouveau texte de l'édito :</label><textarea name='edito' rows='10' required></textarea>";
             $content .= "<input type='submit' name='confirm' value='OK'/></form>";
+            $content .= "<h1>Demandes d'ajout</h1>";
+            $recipesCreation = getWaitingForCreationRecipes();
+            buildRecipeDisplayAllRecipes($content, $recipesCreation, true);
+            $content .= "<h1>Demandes de modification</h1>";
+            $recipesEdition = getWaitingForModificationRecipes();
+            buildRecipeDisplayAllRecipes($content, $recipesEdition, true);
         }
         require('./assets/php/views/accountView.php');
         return;
@@ -174,12 +258,16 @@ function account() {
     require('./assets/php/views/connectionView.php');
 }
 
+function checkCanEditOrDelete($username) {
+    return (isset($_SESSION["userType"]) && strcmp($_SESSION["userType"],"ADMINISTRATEUR") === 0) ||
+    (isset($_SESSION["username"]) && strcmp($_SESSION["username"], $username) === 0 && $_SESSION["userType"] = "EDITEUR");
+}
+
 /* */
 function disconnect() {
     $_SESSION['connected'] = false;
     unset($_SESSION['username']);
     unset($_SESSION['userType']);
-    unset($_SESSION['email']);
     welcome();
 }
 
@@ -193,7 +281,7 @@ function checkIfConnectionValuesExists(&$content) {
     };
     $requiredFieldMissing = false;
     $requiredFieldMissing = match (true) {
-        (empty($_POST['email'])) => $errorMessageFunction($content, "<p>Email manquant !</p>"),
+        (empty($_POST['username'])) => $errorMessageFunction($content, "<p>username manquant !</p>"),
         (empty($_POST['password'])) => $errorMessageFunction($content, "<p>Mot de passe manquant !</p>"),
         default => false,
     };
@@ -202,8 +290,8 @@ function checkIfConnectionValuesExists(&$content) {
         return;
     }*/
     $fieldsMissing = 0;
-    if (empty($_POST['email'])) {
-        $content .= "<p>Email manquant !</p>";
+    if (empty($_POST['username'])) {
+        $content .= "<p>Nom d'utilisateur manquant !</p>";
         $fieldsMissing++;
     }
     if (empty($_POST['password'])) {
@@ -221,16 +309,16 @@ function connectionForm() {
         return;
     }
 
-    $givenEmail = strip_tags($_POST['email']);
+    $givenUsername = strip_tags($_POST['username']);
     $givenPassword = strip_tags($_POST['password']);
 
-    $returnedCredentials = getConnectionCredentials($givenEmail);
+    $returnedCredentials = getConnectionCredentials($givenUsername);
     if (empty($returnedCredentials)) {
-        $content .= "Email incorrect !";
+        $content .= "Nom d'utilisateur incorrect !";
         require('./assets/php/views/connectionView.php');
         return;
     }
-    [$storedUsername, $storedPassword, $storedUserType, $storedEmail] = $returnedCredentials[0];
+    [$storedUsername, $storedPassword, $storedUserType] = $returnedCredentials[0];
     if (!password_verify($givenPassword, $storedPassword)) {
         $content .= "Mot de passe incorrect !";
         require('./assets/php/views/connectionView.php');
@@ -239,16 +327,21 @@ function connectionForm() {
 
     $_SESSION['username'] = $storedUsername;
     $_SESSION['userType'] = $storedUserType;
-    $_SESSION['email'] = $storedEmail;
     $_SESSION['connected'] = true;
 
     $unlogButton = '<button><a href="index.php?action=disconnect">Déconnexion</a></button>';
     $content .= "<p>Connexion réussie. Bienvenue $storedUserType $storedUsername !</p>";
-    if ($_SESSION['userType'] === "ADMINISTRATEUR") {
+    if (strcmp($_SESSION['userType'],"ADMINISTRATEUR") === 0) {
         $content .= "<h1>Modification de l'édito</h1>";
         $content .= "<form id='form_edito' name='edito' method='post' action='index.php?action=editoEdition'>";
         $content .= "<label for='edito'>Entrez le nouveau texte de l'édito :</label><textarea name='edito' rows='10' required></textarea>";
         $content .= "<input type='submit' name='confirm' value='OK'/></form>";
+        $content .= "<h1>Demandes de création</h1>";
+        $recipesCreation = getWaitingForCreationRecipes();
+        buildRecipeDisplayAllRecipes($content, $recipesCreation, true);
+        $content .= "<h1>Demandes de modification</h1>";
+        $recipesEdition = getWaitingForModificationRecipes();
+        buildRecipeDisplayAllRecipes($content, $recipesEdition, true);
     }
     require('./assets/php/views/accountView.php');
 }
@@ -296,8 +389,8 @@ function recipeCreationHandling() {
         require('./assets/php/views/recipeCreationView.php');
         return;
     }
-    $reci_id = createRecipe($title, $desc, $resume, $categorize, $image, $_SESSION["email"]);
-    addRecipesIngredients($reci_id, $ingredients);
+    $reci_id = createRecipe($title, $desc, $resume, $categorize, $image, $_SESSION["username"], strcmp($_SESSION["userType"], "ADMINISTRATEUR") === 0);
+    addRecipesIngredients($reci_id, $ingredients, strcmp($_SESSION["userType"], "ADMINISTRATEUR") === 0);
     getAllRecipes();
 }
 /*recipe modification controller*/
@@ -305,6 +398,8 @@ function recipeEdition($text="") {
     $content = "$text";
     $reci_id = strip_tags($_GET['value']);
     $recipe = getOneRecipe($reci_id);
+    if (count($recipe) === 0) return welcome();
+    if (!checkCanEditOrDelete($recipe[0]["users_nickname"])) return welcome(); 
     $recipeIngredients = getRecipeIngredients($reci_id);
     $title = $recipe[0]["reci_title"];
     $description = $recipe[0]["reci_content"];
@@ -375,8 +470,14 @@ function recipeEditionHandling() {
         $content = "Veuillez mettre au moins un ingrédient s'il vous plaît !";
         return recipeEdition($content);
     }
-    editRecipe($reci_id, $title, $desc, $resume, $categorize, $image, $_SESSION["email"]);
-    editRecipesIngredients($reci_id, $ingredients);
+    if (strcmp($_SESSION["userType"], "ADMINISTRATEUR") === 0) {
+        editRecipe($reci_id, $title, $desc, $resume, $categorize, $image, $_SESSION["username"], true);
+        editRecipesIngredients($reci_id, $ingredients, true);
+    } else {
+        $reci_stash_id = editRecipe($reci_id, $title, $desc, $resume, $categorize, $image, $_SESSION["username"], false);
+        editRecipesIngredients($reci_stash_id, $ingredients, false);
+    }
+    
     getAllRecipes();
 }
 
@@ -384,6 +485,9 @@ function recipeEditionHandling() {
 function recipeDeletion() {
     if (empty($_GET['value'])) getAllRecipes();
     $reci_id = strip_tags($_GET['value']);
+    $recipe = getOneRecipe($reci_id);
+    if (count($recipe) === 0) return getAllRecipes();
+    if (!checkCanEditOrDelete($recipe[0]['users_nickname'])) return getAllRecipes();
     deleteRecipe($reci_id);
     getAllRecipes("<p>La recette a bien été supprimée !</p>");
 }
